cleanup and tests
